# Make QueryBuilder.by_item() faster with a better top k algorithm