--- a/src/reader.rs
+++ b/src/reader.rs
@@ -349,9 +349,7 @@ impl<'t, D: Distance> Reader<'t, D> {
 
         // Get k nearest neighbors
         let k = opt.count.min(nns_distances.len());
-        let top_k = median_based_top_k(k, nns_distances);
-
-        // Normalize distances before return
+        let top_k = median_based_top_k(k, nns_distances, (OrderedFloat(f32::MAX), u32::MAX));
         let mut output = Vec::with_capacity(top_k.len());
         for (OrderedFloat(dist), item) in top_k {
             output.push((item, D::normalized_distance(dist, self.dimensions)));
@@ -550,24 +548,23 @@ pub fn item_leaf<'a, D: Distance>(
     }
 }
 
-pub fn median_based_top_k(
-    k: usize,
-    v: Vec<(OrderedFloat<f32>, u32)>,
-) -> Vec<(OrderedFloat<f32>, u32)> {
+pub fn median_based_top_k<T>(k: usize, v: Vec<T>, mut threshold: T) -> Vec<T>
+where
+    T: Ord + Copy,
+{
     let mut buffer = Vec::with_capacity(2 * k.max(1));
-    let mut threshold = OrderedFloat(f32::MAX);
 
     // prefill with no threshold checks
     let mut v = v.into_iter();
     buffer.extend((&mut v).take(k));
 
     for item in v {
-        if item.0 >= threshold {
+        if item >= threshold {
             continue;
         }
         if buffer.len() == 2 * k {
             let (_, &mut median, _) = buffer.select_nth_unstable(k - 1);
-            threshold = median.0;
+            threshold = median;
             buffer.truncate(k);
         }