# Pull Request
Replaces BinaryHeap with a median-based strategy for calculating top k, reducing complexity from O(n+k\*log(n)) to O(n+k\*log(k))

Not sure if this was something you guys already tried in the past but figured I'd give it a shot since it seems to be relatively low hanging fruit. 

## Related issue
Fixes #72 

## What does this PR do?
This is just a port of a [feature added to tantivy](https://github.com/quickwit-oss/tantivy/pull/2186) back in the day and motivated in [this blog post](https://quickwit.io/blog/top-k-complexity).

Currently arroy finds the top k neighbors by building a BinaryHeap and popping k elements which costs in total O(n) + O(k\*log(n)). The proposed change instead computes the top k by looping over the input and filling a buffer which periodically gets pruned based on the median. The median calculation is done in [O(k)](https://doc.rust-lang.org/std/vec/struct.Vec.html#method.select_nth_unstable) time (but we do it n/k times) and the final sorting is done with QuickSort - O(k\*log(k)). 

When `search_k` >> `count` this feature can yield speedups of ~2% when calculating the nearest neighbours. 

## some benchmarks 
<details>
<summary>Click here</summary>

### dummy data
I wrote some benchmarks [here](https://github.com/nnethercott/arroy_benchmarks/blob/main/benches/reader.rs#L43) using criterion to measure each approach (binary heap vs median-based top k) on some dummy data: 

>*Times for computing top k based on strategy (median-based or binary heap) reported as median/heap for various input sizes, search_k, and nearest neighbors* 

|     | nns=10 | nns=100 | nns=1000 |
|-------|------|-------|-------|
| **search_k=100** | 0.985 µs/1.354 µs| 3.383 µs/4.556 µs| - |
| **search_k=1000** | 2.495 µs/9.703 µs| 12.535 µs/15.577 µs| 45.091 µs/68.696 µs|
| **search_k=10000** | 9.278 µs/88.466 µs | 27.969 µs/100.90 µs| 132.71 µs/184.45 µs|

As k->n the gains diminish as we'd expect. 

### in arroy 
I also benched the performance of `QueryBuilder.by_item()` for different combinations of `nns` and `oversampling`. You can find that [here](https://github.com/nnethercott/arroy_benchmarks/blob/main/benches/reader.rs#L176). In this test setup I used 20k vectors from datacomp small with `n_trees=10`.  Once again the times here are reported as *median / heap*

>*Execution times for QueryBuilder.by_item() for different values of `nns` and `oversampling` for a fixed query*

| | nns=10 | nns=100
-- | -- | --
**oversampling=1** | 265.74 µs / 266.02 µs (−0.28 µs) | 505.00 µs / 503.68 µs (+1.32 µs)
**oversampling=10** | 495.85 µs / 496.62 µs (−0.77 µs) | 2.4891 ms / 2.5164 ms (**−27.3 µs**)
**oversampling=100** | 2.4784 ms / 2.5120 ms (**−33.6 µs**) | 8.6501 ms / 8.7958 ms (**−145.7 µs**)

In extreme cases of oversampling the work of this PR can save a modest fraction of a ms per `by_item` call. 

</details>

## PR checklist
Please check if your PR fulfills the following requirements:
- [x] Does this PR fix an existing issue, or have you listed the changes applied in the PR description (and why they are needed)?
- [x] Have you read the contributing guidelines?
- [x] Have you made sure that the title is accurate and descriptive of the changes?

Thank you so much for contributing to Meilisearch!